+++ src/_begin.js
@@ -1,6 +1,6 @@
 /*!
 Video.js - HTML5 Video Player
-Version 3.0
+Version GENERATED_AT_BUILD
 
 LGPL v3 LICENSE INFO
 This file is part of Video.js. Copyright 2011 Zencoder, Inc.
@@ -22,4 +22,3 @@ along with Video.js.  If not, see <http://www.gnu.org/licenses/>.
 // Self-executing function to prevent global vars and help with minification
 ;(function(window, undefined){
   var document = window.document;
-